errbot-requirements.txt
PyGithub
PyYAML